pyproject.toml
[project]
name = "survey-tag"
version = "0.1.0"
description = "Survey free-text response tagging pipeline"
requires-python = ">=3.13"
dependencies = [
  "bertopic>=0.17.4",
  "numpy>=2.4.4",
  "openai>=2.32.0",
  "pandas>=3.0.2",
  "pyarrow>=24.0.0",
  "python-dotenv>=1.2.2",
  "scikit-learn>=1.8.0",
  "tqdm>=4.67.3",
  "umap-learn>=0.5.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["survey_tag_pipeline"]
